OAuth2Module: enable access tokens to be refreshed concurrently
In the previous implementation, if the refreshAccessToken method
were called while a token refresh was already in progress,
the call would fail and the refresh token would be set to nil,
effectively signing the user out.

This patch introduces a queue which makes sure that any successive
refresh calls will wait for the first call to finish before
returning a result.

Comments in OAuth2Module.swift further explain how it works.